--- src/assets/svgs/index.js
@@ -1,2 +1,0 @@
-export { default as LogoIcon } from './LogoIcon.svg';
-export { default as SearchIcon } from './SearchIcon.svg';
