+++ src/assets/svgs/index.ts
@@ -3,3 +3,5 @@ export { default as LocationIcon } from './LocationIcon.svg';
 export { default as MessageIcon } from './MessageIcon.svg';
 export { default as FileIcon } from './FileIcon.svg';
 export { default as TrashIcon } from './TrashIcon.svg';
+export { default as LogoIcon } from './LogoIcon.svg';
+export { default as SearchIcon } from './SearchIcon.svg';
