+++ src/pages/WritePage.tsx
@@ -1,0 +1,7 @@
+import { WritePage } from 'PageContainer';
+
+const Write = () => {
+  return <WritePage />;
+};
+
+export default Write;
