--- src/theme/index.ts
@@ -1,27 +1,0 @@
-const theme = {
-  input: {
-    width: '804px',
-    border: '1px solid #B8B8B8',
-    borderRadius: '8px',
-    padding: ' 0 0 0 16px',
-    marginBottom: '24px',
-    fontSize: '18px',
-    outline: 'none',
-  },
-  focus: {
-    ':focus': {
-      translate: '0 -2px',
-      transition: 'all 0.2s',
-      boxShadow: '0 4px 10px 0 rgba(0, 0, 0, 0.2)',
-    },
-  },
-  hover: {
-    ':hover': {
-      translate: '0 -2px',
-      transition: 'all 0.2s',
-      boxShadow: '0 4px 10px 0 rgba(0, 0, 0, 0.2)',
-    },
-  },
-};
-
-export default theme;
